pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "custom-action-example"
version = "1.0.0"
description = "Sample DataHub Action for DataHub Cloud (hello_world)"
requires-python = ">=3.10"
dependencies = [
    "acryl-datahub-actions>=1.3",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.entry-points."datahub_actions.action.plugins"]
sample_cloud_action = "custom_action_example.custom_action:CustomAction"
